+++ CHANGELOG.md
@@ -1,10 +1,17 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.1.0] - 2020-09-30
+
+### Fixed
+- Rename private properties
+
+### Added
+- Cache outlets to avoid re-rendering the same content
 
 ## [1.0.0] - 2020-09-29
 
@@ -12,5 +19,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.0] - 2020-09-29
 
-### Added
-- Alpha release
+- Beta release